update tests for validators
